+++ lib/screens/search/search_event.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:meta/meta.dart';
 import 'package:tavern/screens/bloc.dart';
+import 'package:tavern/screens/search/search_qualifiers.dart';
 
 @immutable
 class SearchEvent {}
@@ -17,4 +18,10 @@ class GetSearchHistoryEvent extends SearchEvent {
   final List<String> searchHistory;
 
   GetSearchHistoryEvent({@required this.searchHistory});
+}
+
+class SetSearchQualifierEvent extends SearchEvent {
+  final SearchQualifier searchQualifier;
+
+  SetSearchQualifierEvent(this.searchQualifier);
 }