+++ lib/src/pub_colors.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart' show Color, Colors;
 
 class PubColors {
-  static Color darkColor = Color.fromRGBO(18, 32, 48, 1);
-  static Color darkAccent = Colors.grey[800];
+  static Color darkGunmetal = Color.fromRGBO(18, 32, 48, 1);
+  static Color gunmetal = Color(0xFF30343F);
   static Color searchBarItemsColor = Colors.black45;
   static Color badPackageScore = Color.fromRGBO(187, 36, 0, 1);
   static Color goodPackageScore = Color.fromRGBO(0, 196, 179, 1);
   static Color greatPackageScore = Color.fromRGBO(1, 117, 194, 1);
-  static Color lightBlue = Colors.cyan;
+  static Color turquoiseSurf = Colors.cyan;
+  static Color ghostWhite = Color(0xFFFAFAFF);
 }